Restructure the core.json module to directly expose lodas() and dumps().
 * And ensure consistent use throughout all modules.
 * Fix python-rapdidjson incompatibility.
